requirements.txt
anthropic>=0.40.0
yfinance>=0.2.40
fredapi>=0.5.2
edgartools>=5.0.0
alpaca-py>=0.30.0
finnhub-python>=2.4.0
apscheduler>=3.10.0
jinja2>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
pandas>=2.0.0
sendgrid>=6.10.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
